rename: rename detail components
