+++ src/components/detail/RentButton.js
