+++ src/components/detail/RenteeDetail.js
@@ -10,7 +10,7 @@ import { RiHeart3Fill, RiHeart3Line } from "react-icons/ri";
 import RENTEE_TYPES from "../../constant/RENTEE_TYPES";
 import PopUp from "./PopUp";
 
-function RenteeInfo({
+function RenteeDetail({
   id,
   type,
   thumbnailUrl,
@@ -365,4 +365,4 @@ const AdditionalInfo = styled.div`
   }
 `;
 
-export default RenteeInfo;
+export default RenteeDetail;
